+++ app/views/deceased_settlor/mld5/CountryOfNationalityView.scala.html
@@ -14,7 +14,7 @@
  * limitations under the License.
  *@
 
-@import controllers.deceased_settlor.nonTaxable.routes._
+@import controllers.deceased_settlor.mld5.routes._
 @import utils.InputOption
 @import models.Mode
 @import models.pages.FullName
